requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
matplotlib
pillow
gdown